+++ src/app/photomap/view/mapcom/depth/koreaMap.tsx
