+++ src/app/photomap/view/mapcom/depth/koreaMapImgProto.tsx
