+++ src/app/photomap/view/mapcom/depth/koreaMapProto.tsx
